+++ lib/bashtetikn/html_validator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Bashtetikn
+  # Uses validator provided by W3C to validate HTML
+  # Provides an "generic" interface to an Html Validator
+  class HtmlValidator < HtmlValidatorFromW3C; end
+end